(maint) Update automation for minio image build
Prior to this commit, we were scanning the minio image, but it was left
as a separate action to go out and look for the results.

With this change, we now just update the image daily to ensure it is at
the latest.  Also adds in a CODEOWNERS to be consistent with our
policies and updates the base image to the latest RHEL UBI image.